07_Python_Modulo_7/03_clases_practicas/app_uno/requirements.txt
python-jose[cryptography]
passlib
bcrypt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-multipart
